log output to buffer for test (#685)
