+++ log/ui_test.go
@@ -3,58 +3,23 @@ package log
 import (
 	"testing"
 
-	"bufio"
-	"bytes"
-	"os"
-
 	"github.com/stretchr/testify/assert"
 )
 
-var bufStdOut bytes.Buffer
-var testStdOut *bufio.Writer
-var bufStdErr bytes.Buffer
-var testStdErr *bufio.Writer
-
-func stdToBuf() {
-	testStdOut = bufio.NewWriter(&bufStdOut)
-	stdOut = testStdOut
-	testStdErr = bufio.NewWriter(&bufStdErr)
-	stdErr = testStdErr
-}
-
-func stdToOs() {
-	stdOut = os.Stdout
-	stdErr = os.Stderr
-}
-
-func getStdOut() string {
-	testStdOut.Flush()
-	ret := bufStdOut.String()
-	bufStdOut.Reset()
-	return ret
-}
-
-func getStdErr() string {
-	testStdErr.Flush()
-	ret := bufStdErr.String()
-	bufStdErr.Reset()
-	return ret
-}
-
 func TestMain(m *testing.M) {
-	stdToBuf()
+	OutputToBuf()
 	MainTest(m)
 }
 
 func TestInfo(t *testing.T) {
 	SetDebugVisible(FormatPython)
 	Info("Python")
-	assert.Equal(t, "[+] Python\n", getStdOut())
+	assert.Equal(t, "[+] Python\n", GetStdOut())
 	SetDebugVisible(FormatNone)
 	Info("None")
-	assert.Equal(t, "None\n", getStdOut())
+	assert.Equal(t, "None\n", GetStdOut())
 	Info("None", "Python")
-	assert.Equal(t, "None Python\n", getStdOut())
+	assert.Equal(t, "None Python\n", GetStdOut())
 	SetDebugVisible(1)
 }
 
@@ -62,11 +27,11 @@ func TestLvl(t *testing.T) {
 	SetDebugVisible(1)
 	Info("TestLvl")
 	assert.Equal(t, "I : (                             log.TestLvl:   0) - TestLvl\n",
-		getStdOut())
+		GetStdOut())
 	Print("TestLvl")
 	assert.Equal(t, "I : (                             log.TestLvl:   0) - TestLvl\n",
-		getStdOut())
+		GetStdOut())
 	Warn("TestLvl")
 	assert.Equal(t, "W : (                             log.TestLvl:   0) - TestLvl\n",
-		getStdErr())
+		GetStdErr())
 }
